requirements.txt
Flask
Flask-PyMongo
Flask-Login
Flask-WTF
Flask-Mail
python-dotenv
Werkzeug
pygame
opencv-python

